// src/app/Models/products.ts
export class Product  {
  productId: number;
  productName: string;
  category:string;
  price: number;
  totalprice:number;
  description: string;
  image: string;
  stock:number;
  star:number;
  discount:number;
}

// export const products = [
//   {
//     id: 1,
//     name: 'Phone XL',
//     price: 799,
//     description: 'A large phone with one of the best screens',
//     image: '../assets/images/iphone-XL.jfif'
//   },
//   {
//     id: 2,
//     name: 'Phone Mini',
//     price: 699,
//     description: 'A great phone with one of the best cameras',
//     image: '../assets/images/iphone-Mini.jfif'
//   },
//   {
//     id: 3,
//     name: 'Phone Standard',
//     price: 299,
//     description: '',
//     image: '../assets/images/iphone-13-standard.png'
//   },
//   {
//     id: 4,
//     name: 'Phone XL 2022',
//     price: 899,
//     description: 'A large phone with one of the best screens',
//     image: '../assets/images/iphone-XL.jfif'
//   },
//   {
//     id: 5,
//     name: 'Phone Mini 2022',
//     price: 799,
//     description: 'A great phone with one of the best cameras',
//     image: '../assets/images/iphone-Mini.jfif'
//   },
//   {
//     id: 6,
//     name: 'Phone Standard 2022',
//     price: 399,
//     description: '',
//     image: '../assets/images/iphone-13-standard.png'
//   }
// ];